feat: Add block reduction rogic

+++ backend/app/Savor22b/Model/CookingEquipment.cs
@@ -1,4 +1,5 @@
 namespace Savor22b.Model;
+
 using Libplanet.Assets;
 
 public class KitchenEquipment
@@ -6,7 +7,7 @@ public class KitchenEquipment
     public int ID { get; set; }
     public int KitchenEquipmentCategoryID { get; set; }
     public string Name { get; set; }
-    public double BlockTimeReductionPercent { get; set; }
+    public int BlockTimeReductionPercent { get; set; }
     public string Price { get; set; }
 
     public FungibleAssetValue PriceToFungibleAssetValue()
